+++ src/quicktions.pyx
@@ -1,5 +1,6 @@
 # cython: language_level=3
 ## cython: profile=True
+# cython: freethreading_compatible=True
 
 # Copyright (c) 2001, 2002, 2003, 2004, 2005, 2006, 2007, 2008, 2009, 2010,
 # 2011, 2012, 2013, 2014 Python Software Foundation; All Rights Reserved